api/requirements.txt
mangum
fastapi
langchain
langchain-openai
langchain-core
langchain_community
docx2txt
pypdf
python-dotenv
python-multipart
aiofiles
pinecone
langchain-pinecone
supabase
asyncpg
boto3